+++ .env
@@ -1,11 +1,11 @@
 ### Supabase
-VITE_SUPABASE_URL = "https://oixafyeczqbqnrclpvxo.supabase.co" 
-VITE_SUPABASE_ANON_KEY = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im9peGFmeWVjenFicW5yY2xwdnhvIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MzU0NjcyNDksImV4cCI6MjA1MTA0MzI0OX0.UYC2IC4Rr2xEMoHKuxNvwLfrd7pzlA7wrIZjgwcQzUE"
+VITE_SUPABASE_URL="https://oixafyeczqbqnrclpvxo.supabase.co" 
+VITE_SUPABASE_ANON_KEY="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im9peGFmeWVjenFicW5yY2xwdnhvIiwicm9sZSI6ImFub24iLCJpYXQiOjE3MzU0NjcyNDksImV4cCI6MjA1MTA0MzI0OX0.UYC2IC4Rr2xEMoHKuxNvwLfrd7pzlA7wrIZjgwcQzUE"
 
 ### Stripe
 STRIPE_SECRET_KEY="sk_test_51PC6kIP8Sga8xm9o4Z0U328MPdTFxRSqbaXOJUq6kNPmicpZBlVAAtkDmKD6L63Ib1EGS4Ai44OncEsayRyoue0J00sBNsRHlJ"
-STRIPE_WEBHOOK_SECRET="acct_1PC6kIP8Sga8xm9o"
+STRIPE_WEBHOOK_SECRET="whsec_b39ae5be6cd5b85ee8715cda446b99621f90ddfdd084116a2a495661559770b4"
 
 ### Server
-PORT=5000
+PORT=5001
 CLIENT_URL="http://localhost:5173"
