+++ package.json
@@ -1,7 +1,21 @@
 {
   "name": "starter",
-  "private": true,
   "version": "0.0.0",
+  "proxy": "http://localhost:5173",
+  "private": true,
+  "description": "This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.",
+  "homepage": "https://github.com/rgeron/TempoLabsSWAPV1#readme",
+  "bugs": {
+    "url": "https://github.com/rgeron/TempoLabsSWAPV1/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/rgeron/TempoLabsSWAPV1.git"
+  },
+  "license": "ISC",
+  "author": "",
+  "type": "commonjs",
+  "main": "postcss.config.js",
   "scripts": {
     "dev": "vite",
     "server": "node server.js",
@@ -9,7 +23,8 @@
     "build-no-errors": "tsc ; vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "types:supabase": "npx supabase gen types typescript --project-id $SUPABASE_PROJECT_ID > src/types/supabase.ts"
+    "types:supabase": "npx supabase gen types typescript --project-id $SUPABASE_PROJECT_ID > src/types/supabase.ts",
+    "start": "node server.js"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.6.0",
@@ -40,6 +55,8 @@
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.7",
+    "@stripe/react-stripe-js": "^3.1.1",
+    "@stripe/stripe-js": "^5.5.0",
     "@supabase/supabase-js": "^2.45.6",
     "@types/stripe": "^8.0.417",
     "class-variance-authority": "^0.7.0",
