+++ server.js
@@ -1,8 +1,9 @@
 require("dotenv").config();
-
 const express = require("express");
 const cors = require("cors");
-const Stripe = require("stripe");
+
+// Import the Stripe routes from stripe.js
+const stripeRoutes = require("./stripe.js");
 
 const app = express();
 
@@ -12,92 +13,10 @@ app.use(cors());
 // Parse JSON bodies
 app.use(express.json());
 
-// Initialize Stripe
-console.log("Stripe Secret Key:", process.env.STRIPE_SECRET_KEY);
-
-const stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
-
-// Create a checkout session
-app.post("/api/create-checkout-session", async (req, res) => {
-  try {
-    const { deckId, userId, deckTitle, price } = req.body;
-
-    const session = await stripe.checkout.sessions.create({
-      payment_method_types: ["card"],
-      line_items: [
-        {
-          price_data: {
-            currency: "usd",
-            product_data: {
-              name: deckTitle,
-            },
-            unit_amount: Math.round(price * 100), // Convert to cents
-          },
-          quantity: 1,
-        },
-      ],
-      mode: "payment",
-      success_url: `${process.env.CLIENT_URL}/purchase-success?session_id={CHECKOUT_SESSION_ID}`,
-      cancel_url: `${process.env.CLIENT_URL}/purchase-cancelled`,
-      metadata: {
-        deckId,
-        userId,
-      },
-    });
-
-    res.json({ url: session.url });
-  } catch (error) {
-    console.error("Error creating checkout session:", error);
-    res.status(500).json({ error: "Failed to create checkout session" });
-  }
-});
-
-// Webhook handler
-app.post(
-  "/api/webhook",
-  express.raw({ type: "application/json" }),
-  async (req, res) => {
-    const sig = req.headers["stripe-signature"];
-
-    let event;
-
-    try {
-      event = stripe.webhooks.constructEvent(
-        req.body,
-        sig,
-        process.env.STRIPE_WEBHOOK_SECRET,
-      );
-    } catch (err) {
-      console.error("Webhook signature verification failed:", err);
-      return res.status(400).send(`Webhook Error: ${err.message}`);
-    }
-
-    // Handle the checkout.session.completed event
-    if (event.type === "checkout.session.completed") {
-      const session = event.data.object;
-
-      try {
-        // Extract the deck and user IDs from the session metadata
-        const { deckId, userId } = session.metadata;
-
-        // Update the purchase in your database
-        // await purchaseDeck(userId, deckId);
-
-        console.log(
-          `Successfully processed purchase for deck ${deckId} by user ${userId}`,
-        );
-      } catch (error) {
-        console.error("Error processing purchase:", error);
-        return res.status(500).send("Failed to process purchase");
-      }
-    }
-
-    res.json({ received: true });
-  },
-);
+// Mount the Stripe router on a chosen path prefix, e.g., "/api"
+app.use("/api", stripeRoutes);
 
 const PORT = process.env.PORT || 5000;
-
 app.listen(PORT, () => {
   console.log(`Server running on port ${PORT}`);
 });
