+++ src/App.tsx
@@ -1,19 +1,19 @@
 import { Navigate, Route, Routes } from "react-router-dom";
 import DashboardLayout from "./components/dashboard/DashboardLayout";
-import LandingPage from "./components/landing/LandingPage";
+import CategoryDecks from "./components/dashboard/screens/CategoryDecks";
+import CreatorProfile from "./components/dashboard/screens/CreatorProfile";
+import FollowedCreators from "./components/dashboard/screens/FollowedCreators";
 import Home from "./components/dashboard/screens/HomePage";
 import LikedDecks from "./components/dashboard/screens/LikedDecks";
 import ListedDecks from "./components/dashboard/screens/ListedDecks";
 import PurchasedDecks from "./components/dashboard/screens/PurchasedDecks";
 import SalesAnalytics from "./components/dashboard/screens/SalesAnalytics";
 import SearchDeck from "./components/dashboard/screens/SearchDeck";
-import FollowedCreators from "./components/dashboard/screens/FollowedCreators";
-import CreatorProfile from "./components/dashboard/screens/CreatorProfile";
-import CategoryDecks from "./components/dashboard/screens/CategoryDecks";
-import PurchaseSuccess from "./pages/PurchaseSuccess";
-import PurchaseCancelled from "./pages/PurchaseCancelled";
+import LandingPage from "./components/landing/LandingPage";
 import { Toaster } from "./components/ui/toaster";
 import { AuthProvider } from "./lib/auth";
+import PurchaseCancelled from "./pages/PurchaseCancelled";
+import PurchaseSuccess from "./pages/PurchaseSuccess";
 
 const App = () => {
   return (
@@ -22,8 +22,6 @@ const App = () => {
         {/* Public Routes */}
         <Route path="/" element={<LandingPage />} />
         <Route path="/category/:category" element={<CategoryDecks />} />
-        <Route path="/purchase-success" element={<PurchaseSuccess />} />
-        <Route path="/purchase-cancelled" element={<PurchaseCancelled />} />
 
         {/* App Routes */}
         <Route path="/app" element={<DashboardLayout />}>
@@ -36,6 +34,11 @@ const App = () => {
           <Route path="followed-creators" element={<FollowedCreators />} />
           <Route path="category/:category" element={<CategoryDecks />} />
           <Route path="creator/:creatorId" element={<CreatorProfile />} />
+
+          {/* Make these relative if they are under /app */}
+          <Route path="purchase-success" element={<PurchaseSuccess />} />
+          <Route path="purchase-cancelled" element={<PurchaseCancelled />} />
+
           <Route path="*" element={<Navigate to="home" />} />
         </Route>
       </Routes>
